+++ sdk/core/azure-core/inc/azure/core/credentials/credentials.hpp
@@ -60,6 +60,9 @@ namespace Azure { namespace Core { namespace Credentials {
    * @brief A base type of credential that uses Azure::Core::AccessToken to authenticate requests.
    */
   class TokenCredential {
+  private:
+    std::string m_credentialName;
+
   public:
     /**
      * @brief Gets an authentication token.
@@ -73,13 +76,14 @@ namespace Azure { namespace Core { namespace Credentials {
      */
     virtual AccessToken GetToken(
         TokenRequestContext const& tokenRequestContext,
-        Context const& context) const = 0;
+        Context const& context) const
+        = 0;
 
     /**
      * @brief Gets the name of the credential.
      *
      */
-    virtual std::string GetCredentialName() const;
+    std::string const& GetCredentialName() const { return m_credentialName; }
 
     /**
      * @brief Destructs `%TokenCredential`.
@@ -89,10 +93,24 @@ namespace Azure { namespace Core { namespace Credentials {
 
   protected:
     /**
+     * @brief Constructs an instance of `%TokenCredential`.
+     *
+     * @param credentialName Name of the credential for diagnostic messages.
+     */
+    TokenCredential(std::string credentialName)
+        : m_credentialName(std::move(credentialName.empty() ? "Custom Credential" : credentialName))
+    {
+    }
+
+    /**
      * @brief Constructs a default instance of `%TokenCredential`.
      *
+     * @deprecated Use the constructor with parameter.
      */
-    TokenCredential() {}
+    [[deprecated("Use the construtor with parameter.")]] TokenCredential()
+        : TokenCredential(std::string{})
+    {
+    }
 
   private:
     /**