+++ sdk/core/azure-core/src/credentials/credentials.cpp
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// SPDX-License-Identifier: MIT
-
-#include "azure/core/credentials/credentials.hpp"
-
-using Azure::Core::Credentials::TokenCredential;
-
-std::string TokenCredential::GetCredentialName() const { return "Custom Credential"; }