+++ sdk/identity/azure-identity/inc/azure/identity/default_azure_credential.hpp
@@ -67,12 +67,6 @@ namespace Azure { namespace Identity {
         Core::Credentials::TokenRequestContext const& tokenRequestContext,
         Core::Context const& context) const override;
 
-    /**
-     * @brief Gets the name of the credential.
-     *
-     */
-    std::string GetCredentialName() const override;
-
   private:
     std::shared_ptr<ChainedTokenCredential> m_credentials;
   };