+++ sdk/identity/azure-identity/src/client_certificate_credential.cpp
@@ -74,18 +74,14 @@ void Azure::Identity::_detail::FreePkeyImpl(void* pkey)
   EVP_PKEY_free(static_cast<EVP_PKEY*>(pkey));
 }
 
-std::string ClientCertificateCredential::GetCredentialName() const
-{
-  return "ClientCertificateCredential";
-}
-
 ClientCertificateCredential::ClientCertificateCredential(
     std::string tenantId,
     std::string const& clientId,
     std::string const& clientCertificatePath,
     std::string const& authorityHost,
     TokenCredentialOptions const& options)
-    : m_clientCredentialCore(tenantId, authorityHost),
+    : TokenCredential("ClientCertificateCredential"),
+      m_clientCredentialCore(tenantId, authorityHost),
       m_tokenCredentialImpl(std::make_unique<TokenCredentialImpl>(options)),
       m_requestBody(
           std::string(