+++ sdk/identity/azure-identity/test/ut/token_credential_impl_test.cpp
@@ -34,15 +34,16 @@ public:
       HttpMethod httpMethod,
       Url url,
       TokenCredentialOptions const& options)
-      : m_httpMethod(std::move(httpMethod)), m_url(std::move(url)),
-        m_tokenCredentialImpl(new TokenCredentialImpl(options))
+      : TokenCredential("TokenCredentialImplTester"), m_httpMethod(std::move(httpMethod)),
+        m_url(std::move(url)), m_tokenCredentialImpl(new TokenCredentialImpl(options))
   {
   }
 
   explicit TokenCredentialImplTester(
       std::function<void()> throwingFunction,
       TokenCredentialOptions const& options)
-      : m_throwingFunction(std::move(throwingFunction)),
+      : TokenCredential("TokenCredentialImplTester"),
+        m_throwingFunction(std::move(throwingFunction)),
         m_tokenCredentialImpl(new TokenCredentialImpl(options))
   {
   }
@@ -64,10 +65,21 @@ public:
   }
 };
 
+// Disable deprecation warning
+#if defined(_MSC_VER)
+#pragma warning(push)
+#pragma warning(disable : 4996)
+#endif
+// This credential is needed to test the default behavior when the customer has custom credential
+// they implemented while using earlier versions of the SDK which didn't have a constructor with
+// credentialName.
 class CustomTokenCredential : public TokenCredential {
 public:
   AccessToken GetToken(TokenRequestContext const&, Context const&) const override { return {}; }
 };
+#if defined(_MSC_VER)
+#pragma warning(pop)
+#endif
 
 } // namespace